requirements.txt
aiogram==3.3.0
rembg==2.0.57
Pillow==10.1.0
python-dotenv==1.0.0